Refactor slider's debate & thematics related compontents
