+++ assembl/static2/js/app/components/debate/header/header.jsx
