+++ assembl/static2/js/app/components/debate/navigation/sliderNextArrow.jsx
