+++ assembl/static2/js/app/components/debate/navigation/sliderPrevArrow.jsx
