+++ assembl/static2/js/app/pages/debateThread.jsx
@@ -5,7 +5,7 @@ import { connect } from 'react-redux';
 import { graphql } from 'react-apollo';
 import { Translate } from 'react-redux-i18n';
 import { get } from '../utils/routeMap';
-import Themes from '../components/debate/common/themes';
+import Thematics from '../components/debate/thematics/thematics';
 import Timeline from '../components/debate/navigation/timeline';
 import RootIdeasQuery from '../graphql/RootIdeasQuery.graphql';
 
@@ -38,7 +38,7 @@ const DebateThread = ({ identifier, isNavbarHidden, data, params, children, slug
               <Timeline showNavigation={!isParentRoute} identifier={identifier} />
             </div>
           </section>
-          {isParentRoute && <Themes thematics={thematics} identifier={identifier} />}
+          {isParentRoute && <Thematics thematics={thematics} identifier={identifier} />}
           {!isParentRoute &&
             <section className="debate-section">
               {childrenElm}
